feat: rename init.sh

+++ bin/init/macos.sh
@@ -22,52 +22,51 @@ brew install openssl
 echo 'Cloning dotfiles'
 git clone https://github.com/ebkn/dotfiles ~/dotfiles
 
-echo 'Installing packages from homebrew'
-brew bundle --file="~/dotfiles/brewfiles/Brewfile-shell"
+echo 'Setup shell...'
+chsh -s /usr/local/bin/zsh
 
-# install zplugin
+echo 'Starting zsh'
+zsh
+
+echo 'Installing zplugin'
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/zdharma/zplugin/master/doc/install.sh)"
+[ -f ~/.zshrc ] && mv ~/.zshrc ~/backup/
+[ -f ~/.zshenv ] && mv ~/.zshenv ~/backup/
+ln -s ~/dotfiles/.zshrc ~
+ln -s ~/dotfiles/.zshenv ~
 
-mv ~/.tmux.conf ~/backup/
-mv ~/.zshrc ~/backup/
-mv ~/.zshenv ~/backup/
-mv ~/.bash_profile ~/backup/
-mv ~/.bashrc ~/backup/
-mv ~/.vimrc ~/backup/
-mv ~/.vim ~/backup/
-
-ln -s ~/dotfiles/.gitignore_global ~
-ln -s ~/dotfiles/.vim ~
-
+echo 'Installing dein'
 curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh >  ~/installer.sh
 sh ~/installer.sh ~/.cache/dein
 rm ~/installer.sh
 
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-~/.fzf/install
+echo 'Installing shell packages'
+brew bundle --file="~/dotfiles/brewfiles/Brewfile-shell"
 
-echo 'Setup shell...'
-chsh -s /usr/local/bin/zsh
-echo 'please restart teminal to apply changes'
+[ -f ~/.tmux.conf ] && mv ~/.tmux.conf ~/backup/
+[ -f ~/.bash_profile ] && mv ~/.bash_profile ~/backup/
+[ -f ~/.bashrc ] && mv ~/.bashrc ~/backup/
+[ -f ~/.vimrc ] && mv ~/.vimrc ~/backup/
+[ -d ~/.vim ] && mv ~/.vim ~/backup/
+[ -f ~/.tigrc ] && mv ~/.tigrc ~/backup/
 
-# load shell settings
+ln -s ~/dotfiles/.gitignore_global ~
 ln -s ~/dotfiles/.bash_profile ~
 ln -s ~/dotfiles/.bashrc ~
-ln -s ~/dotfiles/.zshrc ~
-ln -s ~/dotfiles/.zshenv ~
 ln -s ~/dotfiles/.tmux-conf ~
 ln -s ~/dotfiles/.vimrc ~
 ln -s ~/dotfiles/.vim ~
 ln -s ~/dotfiles/.tigrc ~
 
-source ~/.bash_profile
 source ~/.zshrc
+
+echo 'Starting tmux'
+tmux
 tmux source-file ~/.tmux.conf
 
-echo 'Installing nvm and node'
-# brew install nodenv
-curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.11/install.sh | zsh
-nvm install node
+echo 'Installing fzf'
+git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+~/.fzf/install
 
 echo 'Installing languages from homebrew'
 brew bundle --file='brewfiles/Brewfile-lang'
@@ -79,8 +78,8 @@ ln -s ~/dotfiles/tsconfig.json ~
 echo 'Installing apps by Homebrew-Cask..'
 brew bundle --file='~/brewfiles/Brewfile-cask'
 
-mv ~/.config/alacritty ~/backup/
+[ -d ~/.config/alacritty ] && mv ~/.config/alacritty ~/backup/
 ln -s ~/dotfiles/alacritty/.alacritty.mac.yml ~/.alacritty.yml
 
-mv .gitconfig ~/backup/
+[ -f ~/.gitconfig ] && mv ~/.gitconfig ~/backup/
 ln -s ~/dotfiles/.gitconfig ~/.gitconfig
